Feat: Refactoring Rgbd_images; Refactoring KinematicModelChecker; Few changes on markers and processing
